product type delete and update services

+++ src/main/java/com/crudoperation/jw/repo/ProductTypeRepository.java
@@ -3,5 +3,5 @@ package com.crudoperation.jw.repo;
 import com.crudoperation.jw.model.ProductType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductTypeRepository extends JpaRepository<ProductType, Long> {
+public interface ProductTypeRepository extends JpaRepository<ProductType, Integer> {
 }
